storstädning
delat upp koden i olika filer:
model, provider och api.

städat bort alla prints och kompileringsfel

verkar stabilt
