+++ good_morning/lib/ui/main_navigation/home_page.dart
@@ -14,7 +14,7 @@ import 'package:good_morning/ui/daily_film/daily_film_page.dart';
 import 'package:good_morning/ui/daily_traffic/daily_traffic.ui.dart';
 import '../../utils/filter_model.dart';
 import 'onboarding.dart';
-import 'package:good_morning/utils/daily_history.dart';
+import 'package:good_morning/utils/daily_history/daily_history_provider.dart';
 import 'package:good_morning/utils/weather.dart';
 
 class HomePage extends StatefulWidget {
